HYDRA-372 fix streaming source api
Fixing the interface so that it takes a Hydrator context instead
of only the Spark context. This gives sources access to additional
information, such as metrics, the stage name, plugins, etc.

Along the way, refactoring common stage related methods into a
StageContext interface that everything inherits from. This changes
the TransformContext so that it does not directly extend the CDAP
PluginContext, as that interface may contain methods that dont
make sense for Hydrator, such as the method that takes a
MacroEvaluator to instantiate a plugin.
